Quality increase!
I've rewritten my scripts (or even replaced most of them really) with higher quality, better coded versions!

autoPlant still maintains a list (now a dictionary) of all servers, automatically breaking/hacking new ones for you through a run as your hacker level and number of port breakers increase.
hackStarter is an initializer for autoHack, copying it where it needs to go and sending it necessary info.
autoHack is a fairly simple auto-hacking loop.
serverListBuilder builds a dictionary of all game servers, along with convenient info like max money and RAM for each (to save you the RAM elsewhere), and exports it via a port (default 20).
functionLibrary is just a simple library of functions that the other scripts make use of!